+++ interceptor/middleware/responsewriter_interface.go
@@ -4,9 +4,10 @@ import (
 	"net/http"
 )
 
-// HijackerResponseWriter combines http.ResponseWriter and http.Hijacker interfaces
-// This is used for generating mocks that implement both interfaces
+// HijackerResponseWriter combines http.ResponseWriter, http.Hijacker and EnableFullDuplex interfaces
+// This is used for generating mocks that implement all interfaces
 type HijackerResponseWriter interface {
 	http.ResponseWriter
 	http.Hijacker
+	EnableFullDuplex() error
 }